Provide start_order and stop_order for devices, to specify what order BLACS should transition them to buffered mode and static mode respectively. Allows any integer, default is zero. If specified on a device without a BLACS_connection attribute, a TypeError is raised during compilation. start_order and stop_order are saved as device_properties, and thus can vary from shot-to-shot without a connection table recompilation.
At a later date (labscript 3?), devices connected directly to BLACS will
probably be a subclass of a BLACSDevice mixin class or something like that,
at which point we can make only these classes have the keyword arguments in the
first place. For the moment however this seems like a workable implementation.