Move trigger validity checks into TriggerableDevice from the Camera class.
They are applicable to all triggerable devices, not just the Camera, so
it avoids code duplication to have them here. Devices requesting triggers at
particular times should call self.trigger() instead of self.trigger_device.trigger(),
in order to take advantage of the checks.

This was motivated by me implementing a new camera class, this way I will
not be duplicating code.

I'll make a separate change to the Camera class in labscript_devices,
removing the code from there and depending on this version of labscript.
It is harmless having the checks be in both places, the checks in
labscript will not find any triggers to throw errors about until device classes
are updated to call self.trigger() instead of self.trigger_device.trigger().

Spurious errors could be thrown if some devices called self.trigger() and others
called self.trigger_device.trigger(), so all the devices in labscript_devices
will be updated at once (I think it's just the camera, but I'll check).