Allow setting marker colors using strings such as 'red' or '#ff8g80'. If no colour is set, save the color as (-1, -1, -1), to distinguish 'no colour set' from black. This way, actually setting the colour to black can be done, whereas the BLACS progress bar can fall back to default colours if no colour has been set (right now it uses default colours if it encounters black, precluding one from actually using black).
If BLACS and runviewer are not updated after this change, they will encounter
colours that are (-1, -1, -1). They will not break, but they will print a
warning from Qt and will treat the colour as black. So this is fairly backward
compatible - the only visible consequence is the BLACS progress bar will be black
instead of blue for markers with no colour set, until BLACS is updated to be
aware of the change.

Corresponding changes to BLACS and runviewer will make them aware of the
distinction between black and 'unset'.