Make wait monitors optional.
It will be up to individual devices whether to raise an exception if they are
asked to do analog input and require a wait monitor when one is not present,
or if they require a wait monitor for any other purpose.

If using a wait monitor, make the use of a timeout device optional.

Add a detailed docstring for WaitMonitor